Switch to docker buildx to facilitate building for (or on) ARM64.
Note that due to #304 the frontend nodebuilder stage is hard-coded to always run with --platform=linux/amd64, therefore building on arm64 _requires_ QEMU to be properly set up and the active buildx builder to be capable of building amd64 as well as arm64.  This is the default setup with Docker Desktop for Mac on Apple Silicon but may require extra setup on e.g. AWS Graviton instances.