Support fromDocument property being a single string
The structured array or object forms of fromDocument options work well when documents are supplied in JSON but are tricky to map effectively when uploading documents as CSV.  We now also support the case where the property pointed to by fromDocument is a single string.  By default this is assumed to be a semicolon-separated list of options, where each option is of the form value=label, i.e.

    {"options":"orange=Orange; kiwi=Kiwi fruit"}

is treated as if it were

    {"options":[
      {"value": "orange", "label": "Orange},
      {"value": "kiwi", "label": "Kiwi fruit"}
    ]}

The ';' and '=' separators are configurable by properties alongside the "fromDocument", and can be any string (not just single characters), e.g.

    {"fromDocument": "options", "separator":"##", "valueLabelSeparator": ":"}

with document

    {"options":"option1: Option 1##option2: Option 2"}