Initial attempt at implementing #164 (conditional annotation)
Added an "if" option to all annotation widgets, whose value is an expression that has access to the document data _and_ the current state of annotationOutput.  Widgets are only rendered if their if expression evaluates to a truthy value.  The conditionals are re-evaluated every time the annotation data changes, so widgets can appear or disappear depending on the current state.