New approach - rather than allowing functions by default and then trying to play whack-a-mole with the ones we want to forbid as "dangerous", instead forbid _all_ function calls by default and add special syntax for the cases that we do want to allow.
Allowing things like Object.assign is a big hole as it allows "if" expressions to pollute the global Array.prototype, etc., allowing arbitrary method calls lets the expression call mutating methods like someArray.push(), ...  Determining which functions are "safe" and which aren't is asking for trouble.  Given the point of these expressions is to determine whether or not to display other widgets, they shouldn't really need to use functions much at all.  The main classes of expressions we need to support are:

- simple single-value binary operator comparisons - annotation.something === 'value', annotation.confidence < 4, etc.
- universal and existential quantifiers over arrays of checkbox values
  - any(cb in annotation.someCheckbox, cb === 'other')
  - all(cb in annotation.someCheckbox, cb > 3)
- or over values from the document
  - any(v in document.validChoices, annotation.choice === v)

In addition regular expression matching of some kind is useful for conditions over free-text strings, and in JS this normally requires a function call (/regex/.test(str)) - we borrow syntax from Perl to convert this into a binary operation instead of a call (str =~ /regex/).

With this new syntax in place we can forbid _all_ other function and method calls.  This, along with also forbidding access to the __ob__ property (that Vue adds to all observable objects, and that has a chain of properties leading to the global "window"), should give the expressions sufficient power to be useful but not permit them to modify the document/annotation or read anything outside of the context data itself.