In test_annotations_per_doc_not_enforced_for_training_or_test, instead of adding one annotator and having them do all the training, test and annotation docs, then adding the next annotator, etc. we now add all the annotators, then have them all do the training, then all do the test, then all do the annotations documents.
This is because it is possible for just five of the six annotators to between them triple-annotate the entire corpus (each annotator is allowed up to 12 documents, and 60 annotations in total would fill the corpus), causing an error when the test attempts to add a sixth annotator to the already-completed project.  Adding all the annotators up front before anyone starts working is a more realistic test.